+++ ReviewService/src/main/java/org/app/reviewservice/service/ReviewService.java
@@ -1,16 +1,21 @@
 package org.app.reviewservice.service;
 
+import jakarta.validation.Valid;
 import org.app.reviewservice.dto.ReviewDTO;
 import org.app.reviewservice.entity.Review;
+import org.app.reviewservice.exception.ConflictException;
 import org.app.reviewservice.mapper.ReviewMapper;
 import org.app.reviewservice.repository.ReviewRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.validation.annotation.Validated;
 
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
+@Validated
 @Service
 public class ReviewService {
   private final ReviewRepository reviewRepository;
@@ -27,18 +32,28 @@ public class ReviewService {
             .toList();
   }
 
-  public Optional<ReviewDTO> getReviewById(Long id) {
+  public ReviewDTO getReviewById(Long id) {
     return reviewRepository.findById(id)
-            .map(reviewMapper::toDTO);
+            .map(reviewMapper::toDTO)
+            .orElseThrow(() -> new NoSuchElementException("Review with ID " + id + " not found"));
   }
 
-  public ReviewDTO saveReview(ReviewDTO reviewDTO) {
+  public ReviewDTO saveReview(@Valid ReviewDTO reviewDTO) {
+    boolean exists = reviewRepository.existsByUserIdAndEventId(reviewDTO.getUserId(), reviewDTO.getEventId());
+
+    if (exists) {
+      throw new ConflictException("User has already submitted a review for this event");
+    }
+
     Review review = reviewMapper.toEntity(reviewDTO);
     review.setTimestamp(Timestamp.valueOf(LocalDateTime.now())); // Set current timestamp
     return reviewMapper.toDTO(reviewRepository.save(review));
   }
 
   public void deleteReview(Long id) {
+    if (!reviewRepository.existsById(id)) {
+      throw new NoSuchElementException("Review with ID " + id + " does not exist");
+    }
     reviewRepository.deleteById(id);
   }
 }
